+++ packages/@office-iss/react-native-win32/src/Libraries/Components/TextInput/Win32TextInputNativeComponent.js
@@ -1,0 +1,23 @@
+/**
+ * @flow
+ * @format
+ */
+
+import type {HostComponent} from '../../Renderer/shims/ReactNativeTypes';
+
+import requireNativeComponent from '../../ReactNative/requireNativeComponent';
+import codegenNativeCommands from '../../Utilities/codegenNativeCommands';
+import type {TextInputNativeCommands} from './TextInputNativeCommands';
+type NativeType = HostComponent<mixed>;
+
+type NativeCommands = TextInputNativeCommands<NativeType>;
+
+export const Commands: NativeCommands = codegenNativeCommands<NativeCommands>({
+  supportedCommands: ['focus', 'blur', 'setTextAndSelection'],
+});
+
+const WindowsTextInputComponent: NativeType =
+  requireNativeComponent<mixed>('RCTTextInput');
+
+export default WindowsTextInputComponent;
+// [Windows]